+++ dejafu/CHANGELOG.rst
@@ -6,8 +6,11 @@ standard Haskell versioning scheme.
 
 .. _PVP: https://pvp.haskell.org/
 
-unreleased
-----------
+1.7.0.0 (2018-06-03)
+--------------------
+
+* Git: :tag:`dejafu-1.7.0.0`
+* Hackage: :hackage:`dejafu-1.7.0.0`
 
 Changed
 ~~~~~~~